// react-obento/src/api/bento-service/models/BentoRecord.ts
/* tslint:disable */
/* eslint-disable */
/**
 * obento-service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Who } from './Who';
import {
    WhoFromJSON,
    WhoFromJSONTyped,
    WhoToJSON,
    WhoToJSONTyped,
} from './Who';

/**
 * 
 * @export
 * @interface BentoRecord
 */
export interface BentoRecord {
    /**
     * 
     * @type {number}
     * @memberof BentoRecord
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof BentoRecord
     */
    userId?: string;
    /**
     * 
     * @type {Date}
     * @memberof BentoRecord
     */
    date?: Date;
    /**
     * 
     * @type {Who}
     * @memberof BentoRecord
     */
    who?: Who;
}



/**
 * Check if a given object implements the BentoRecord interface.
 */
export function instanceOfBentoRecord(value: object): value is BentoRecord {
    return true;
}

export function BentoRecordFromJSON(json: any): BentoRecord {
    return BentoRecordFromJSONTyped(json, false);
}

export function BentoRecordFromJSONTyped(json: any, ignoreDiscriminator: boolean): BentoRecord {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'userId': json['userId'] == null ? undefined : json['userId'],
        'date': json['date'] == null ? undefined : (new Date(json['date'])),
        'who': json['who'] == null ? undefined : WhoFromJSON(json['who']),
    };
}

export function BentoRecordToJSON(json: any): BentoRecord {
    return BentoRecordToJSONTyped(json, false);
}

export function BentoRecordToJSONTyped(value?: BentoRecord | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'userId': value['userId'],
        'date': value['date'] == null ? undefined : ((value['date']).toISOString().substring(0,10)),
        'who': WhoToJSON(value['who']),
    };
}