// react-obento/src/api/bento-service/models/Who.ts
/* tslint:disable */
/* eslint-disable */
/**
 * obento-service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const Who = {
    Self: 'self',
    Mom: 'mom',
    Buy: 'buy'
} as const;
export type Who = typeof Who[keyof typeof Who];


export function instanceOfWho(value: any): boolean {
    for (const key in Who) {
        if (Object.prototype.hasOwnProperty.call(Who, key)) {
            if (Who[key as keyof typeof Who] === value) {
                return true;
            }
        }
    }
    return false;
}

export function WhoFromJSON(json: any): Who {
    return WhoFromJSONTyped(json, false);
}

export function WhoFromJSONTyped(json: any, ignoreDiscriminator: boolean): Who {
    return json as Who;
}

export function WhoToJSON(value?: Who | null): any {
    return value as any;
}

export function WhoToJSONTyped(value: any, ignoreDiscriminator: boolean): Who {
    return value as Who;
}